+++ WEB(FE)/src/components/common/Drawer.jsx
@@ -2,7 +2,7 @@ import { Link } from 'react-router-dom';
 import tw from 'tailwind-styled-components';
 
 const DrawerSide = tw.div`
-  drawer-sideS
+  drawer-side
 `;
 
 const Drawer = ({ closeOverlay }) => {
@@ -10,16 +10,38 @@ const Drawer = ({ closeOverlay }) => {
     <DrawerSide>
       <label htmlFor='side-menu' className='drawer-overlay'></label>
       <ul className=' menu w-60 sm:w-80 p-4 overflow-y-auto bg-white dark:bg-base-100'>
-        <li>
-          <input type='search' className='input input-bordered' />
+        <li className='mt-2'>
+          <input
+            type='search'
+            placeholder='SEARCH'
+            className='input input-bordered'
+          />
         </li>
-        <li>
+        <li className='mt-2'>
           <Link
-            className='text-gray-700 active:text-white dark:text-white'
-            to='/'
+            className='bg-gray-300 hover:bg-gray-500 text-gray-700 active:text-white dark:text-white'
+            to='login'
             onClick={closeOverlay}
           >
-            "menu.title"
+            로그인
+          </Link>
+        </li>
+        <li className='mt-2'>
+          <Link
+            className='bg-gray-300 hover:bg-gray-500 text-gray-700 active:text-white dark:text-white'
+            to='/map'
+            onClick={closeOverlay}
+          >
+            지도
+          </Link>
+        </li>
+        <li className='mt-2'>
+          <Link
+            className='bg-gray-300 hover:bg-gray-500 text-gray-700 active:text-white dark:text-white'
+            to='/posts'
+            onClick={closeOverlay}
+          >
+            게시판
           </Link>
         </li>
       </ul>
