+++ WEB(FE)/src/pages/Mainpage.jsx
@@ -1,30 +1,14 @@
-import { useRef } from 'react';
-import DrawerContainer from '../containers/common/DrawerContainer';
 import FooterContainer from '../containers/common/FooterContainer';
 import HeaderContainer from '../containers/common/HeaderContainer';
 import ItemsContainer from '../containers/intro/ItemsContainer';
 
 const Mainpage = () => {
-  const $hamburger = useRef(null);
-  const $pages = useRef(null);
-  const closeOverlay = () => {
-    $hamburger.current.click();
-  };
   return (
-    <div className='drawer'>
-      <input
-        type='checkbox'
-        id='side-menu'
-        className='drawer-toggle'
-        ref={$hamburger}
-      />
-      <section className='drawer-content relative' ref={$pages}>
-        <HeaderContainer />
-        <ItemsContainer />
-        <FooterContainer />
-      </section>
-      <DrawerContainer closeOverlay={closeOverlay} />
-    </div>
+    <>
+      <HeaderContainer />
+      <ItemsContainer />
+      <FooterContainer />
+    </>
   );
 };
 export default Mainpage;
